Split the route monolith; add per-device presence and --on targeting
- server/routes.ts (1.6k lines) becomes routers per concern: auth/devices,
  artifacts, actions, display, integrations (Nexlayer/chat/PDF proxies
  quarantined); shared render helpers move to server/render.ts
- SSE connections are tagged with a delivery target (device session id or
  'local'); directed events deliver to one screen only
- Presence is session-keyed (server/presence.ts): /display/status returns the
  per-device list, /display/presence files under the caller's device
- surface open/notify gain --on <device> (case-insensitive prefix match;
  ambiguous or unknown names error with the device list)
- surface devices shows live connection state and what each display is viewing
