+++ server/routes/actions.ts
@@ -1,0 +1,146 @@
+import { Router } from "express";
+import { ackAction, createAction, getDb, getPendingActions } from "../db.js";
+import { getArtifact } from "../artifacts.js";
+import { broadcastGlobal, broadcastToSurface } from "../sse.js";
+import { requireSystem } from "./helpers.js";
+
+export const actionsRouter = Router();
+
+// Optional webhook fan-out for surface actions. Set SURFACE_WEBHOOK_URL and
+// SURFACE_WEBHOOK_TOKEN to wake an external agent gateway when users interact
+// with surfaces. OPENCLAW_GATEWAY_URL / OPENCLAW_HOOKS_TOKEN are kept as
+// legacy aliases for older configs.
+const WEBHOOK_URL = process.env.SURFACE_WEBHOOK_URL || process.env.OPENCLAW_GATEWAY_URL;
+const WEBHOOK_TOKEN = process.env.SURFACE_WEBHOOK_TOKEN || process.env.OPENCLAW_HOOKS_TOKEN;
+const WEBHOOK_PATH = process.env.SURFACE_WEBHOOK_PATH || "/hooks/agent";
+
+// Suppress webhook-failure notifications to at most one per minute so a broken
+// webhook doesn't flood the display with toasts.
+let lastWebhookNotifyAt = 0;
+const WEBHOOK_NOTIFY_THROTTLE_MS = 60_000;
+
+function notifyWebhookFailure(reason: string) {
+  const now = Date.now();
+  if (now - lastWebhookNotifyAt < WEBHOOK_NOTIFY_THROTTLE_MS) return;
+  lastWebhookNotifyAt = now;
+  broadcastGlobal("display_notify", {
+    text: `Webhook fan-out failed: ${reason}`,
+    duration: 5000,
+    style: "warning",
+  });
+}
+
+async function fanOutWebhook(payload: {
+  surface_id: string;
+  surface_title: string;
+  action: string;
+  data: unknown;
+  created_at: string;
+}) {
+  if (!WEBHOOK_URL || !WEBHOOK_TOKEN) return;
+  try {
+    const res = await fetch(`${WEBHOOK_URL}${WEBHOOK_PATH}`, {
+      method: "POST",
+      headers: {
+        Authorization: `Bearer ${WEBHOOK_TOKEN}`,
+        "Content-Type": "application/json",
+      },
+      body: JSON.stringify({ type: "surface_action", ...payload }),
+    });
+    if (!res.ok) {
+      const body = await res.text();
+      console.error(`[webhook] ${WEBHOOK_URL}${WEBHOOK_PATH} returned ${res.status}: ${body}`);
+      notifyWebhookFailure(`${res.status} ${res.statusText}`);
+    }
+  } catch (err: any) {
+    console.error(`[webhook] dispatch failed:`, err);
+    notifyWebhookFailure(err?.message || "network error");
+  }
+}
+
+// Display posts a user action (iframe postMessage → PWA → here).
+actionsRouter.post("/artifacts/:id/actions", (req, res) => {
+  const artifact = getArtifact(getDb(), req.params.id);
+  if (!artifact) {
+    res.status(404).json({ error: "Artifact not found" });
+    return;
+  }
+  const { action, data } = req.body;
+  if (!action) {
+    res.status(400).json({ error: "action is required" });
+    return;
+  }
+  const act = createAction({ surface_id: req.params.id, action, data });
+  fanOutWebhook({
+    surface_id: req.params.id,
+    surface_title: artifact.title,
+    action: act.action,
+    data: data ?? {},
+    created_at: act.created_at,
+  });
+  broadcastGlobal("surface_action", {
+    id: act.id,
+    surface_id: req.params.id,
+    surface_title: artifact.title,
+    action: act.action,
+    data: act.data,
+    created_at: act.created_at,
+  });
+  res.status(201).json(act);
+});
+
+// Agent reads pending actions — the inbox belongs to the agent plane; a device
+// must never drain it.
+actionsRouter.get("/actions", (req: any, res) => {
+  if (!requireSystem(req, res)) return;
+  res.json(getPendingActions());
+});
+
+actionsRouter.get("/artifacts/:id/actions", (req: any, res) => {
+  if (!requireSystem(req, res)) return;
+  res.json(getPendingActions(req.params.id));
+});
+
+// Agent acknowledges an action
+actionsRouter.post("/actions/:id/ack", (req: any, res) => {
+  if (!requireSystem(req, res)) return;
+  const acked = ackAction(req.params.id);
+  if (!acked) {
+    res.status(404).json({ error: "Action not found" });
+    return;
+  }
+  res.json({ acknowledged: true });
+});
+
+// Agent replies to a surface (shown as toast in the PWA)
+actionsRouter.post("/artifacts/:id/reply", (req: any, res) => {
+  if (!requireSystem(req, res)) return;
+  if (!getArtifact(getDb(), req.params.id)) {
+    res.status(404).json({ error: "Artifact not found" });
+    return;
+  }
+  const { text } = req.body;
+  if (!text) {
+    res.status(400).json({ error: "text is required" });
+    return;
+  }
+  broadcastToSurface(req.params.id, "agent_reply", { text });
+  broadcastGlobal("agent_reply", { surface_id: req.params.id, text });
+  res.json({ sent: true });
+});
+
+// Execute JS in a surface iframe — code execution, system plane only.
+actionsRouter.post("/artifacts/:id/exec", (req: any, res) => {
+  if (!requireSystem(req, res)) return;
+  if (!getArtifact(getDb(), req.params.id)) {
+    res.status(404).json({ error: "Artifact not found" });
+    return;
+  }
+  const { js } = req.body;
+  if (!js) {
+    res.status(400).json({ error: "js is required" });
+    return;
+  }
+  broadcastToSurface(req.params.id, "surface_exec", { js });
+  res.json({ executed: true });
+});
